feat: normalize caudalDiseno variants
